WARNING - cellular API change: uCellInfoGetTimeUtc() now returns an int64_t. (#277)
With this commit uCellInfoGetTimeUtc() is brought into line with all of the other UTC time functions in ubxlib, returning an int64_t rather than an int32_t.

+++ cell/api/u_cell_info.h
@@ -273,7 +273,7 @@ int32_t uCellInfoGetFirmwareVersionStr(int32_t cellHandle,
  * @return            on success the Unix UTC time, else negative
  *                    error code.
  */
-int32_t uCellInfoGetTimeUtc(int32_t cellHandle);
+int64_t uCellInfoGetTimeUtc(int32_t cellHandle);
 
 #ifdef __cplusplus
 }
